@@ -1,10 +1,8 @@
 package com.giffing.wicket.spring.boot.example.web;
 
-import java.util.Arrays;
-import java.util.HashSet;
 import java.util.Set;
 
-import javax.servlet.SessionTrackingMode;
+import jakarta.servlet.SessionTrackingMode;
 
 import org.apache.wicket.Application;
 import org.apache.wicket.protocol.http.mock.MockServletContext;
@@ -17,12 +15,12 @@ public class WicketMockServletContext extends MockServletContext {
 	
 	@Override
 	public Set<SessionTrackingMode> getDefaultSessionTrackingModes() {
-		return new HashSet<>(Arrays.asList( SessionTrackingMode.COOKIE ));
+		return Set.of(SessionTrackingMode.COOKIE);
 	}
 
 	@Override
 	public Set<SessionTrackingMode> getEffectiveSessionTrackingModes() {
-		return new HashSet<>(Arrays.asList( SessionTrackingMode.COOKIE ));
+		return Set.of(SessionTrackingMode.COOKIE);
 	}
 
 }
