@@ -1,24 +1,19 @@
 package com.giffing.wicket.spring.boot.example.web;
 
+import com.giffing.wicket.spring.boot.example.WicketApplication;
+import com.giffing.wicket.spring.boot.example.web.pages.login.LoginPage;
+import com.giffing.wicket.spring.boot.starter.configuration.extensions.external.spring.security.SecureWebSession;
+import com.giffing.wicket.spring.boot.starter.web.servlet.websocket.WebSocketMessageBroadcaster;
 import org.apache.wicket.protocol.http.WebApplication;
 import org.apache.wicket.util.tester.FormTester;
 import org.apache.wicket.util.tester.WicketTester;
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Disabled;
-import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mockito;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Primary;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.test.context.junit.jupiter.SpringExtension;
-
-import com.giffing.wicket.spring.boot.example.WicketApplication;
-import com.giffing.wicket.spring.boot.example.web.pages.login.LoginPage;
-import com.giffing.wicket.spring.boot.starter.configuration.extensions.external.spring.security.SecureWebSession;
-import com.giffing.wicket.spring.boot.starter.web.servlet.websocket.WebSocketMessageBroadcaster;
 
 /**
  * Test class for initialize Wicket & Spring Boot only in the web package. All
@@ -27,11 +22,8 @@ import com.giffing.wicket.spring.boot.starter.web.servlet.websocket.WebSocketMes
  * @author Marc Giffing
  *
  */
-@ExtendWith(SpringExtension.class)
 @SpringBootTest(classes = WicketApplication.class)
-@EnableWebSecurity
-@Disabled
-public class WicketBaseIntTest {
+public abstract class WicketBaseIntTest {
 
 	private static final String USERNAME = "admin";
 	private static final String PASSWORD = "admin";
@@ -49,8 +41,6 @@ public class WicketBaseIntTest {
 			return Mockito.mock(WebSocketMessageBroadcaster.class);
 		}
 	}
-
-	
 
 	@BeforeEach
 	public void setUp() {
