@@ -5,7 +5,6 @@ import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Import;
 
-import com.giffing.wicket.spring.boot.starter.app.classscanner.ClassCandidateScanner;
 import com.giffing.wicket.spring.boot.starter.app.classscanner.ClassCandidateScannerConfiguration;
 import com.giffing.wicket.spring.boot.starter.app.verifier.WicketDependencyVersionChecker;
 import com.giffing.wicket.spring.boot.starter.configuration.CustomAnnotationBeanNameGenerator;
@@ -15,7 +14,7 @@ import com.giffing.wicket.spring.boot.starter.web.WicketWebInitializer;
 
 /**
  * The main starter configuration class which will be called by spring.
- * The class is configured in META-INF/spring.factories
+ * The class is configured in META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
  * 
  * @author Marc Giffing
  *
